+++ MUMENT/MUMENT/Sources/Scenes/Home/SongDetail/AllMumentsSectionHeader.swift
@@ -9,43 +9,6 @@ import UIKit
 import SnapKit
 import Then
 
-class OrderingButton: UIButton {
-    private var buttonText = ""
-    private func setConfiguration(){
-        var configuration = UIButton.Configuration.plain()
-        configuration.imagePlacement = .trailing
-        configuration.buttonSize = .small
-        configuration.baseBackgroundColor = UIColor.mBgwhite
-        self.configuration = configuration
-    }
-    
-    private func setTextStyle(){
-        self.setAttributedTitle(NSAttributedString(string: buttonText, attributes: [
-            .font: UIFont.mumentC1R12,
-            .foregroundColor: UIColor.mPurple1
-        ]), for: .selected)
-        self.setAttributedTitle(NSAttributedString(string: buttonText, attributes: [
-            .font: UIFont.mumentC1R12,
-            .foregroundColor: UIColor.mGray1
-        ]), for: .normal)
-    }
-    
-    required init(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)!
-    }
-    
-    override init(frame: CGRect){
-        super.init(frame: frame)
-    }
-    
-    init(_ text:String) {
-        super.init(frame:CGRect.zero)
-        buttonText = text
-        setConfiguration()
-        setTextStyle()
-    }
-}
-
 class AllMumentsSectionHeader: UITableViewHeaderFooterView {
     
     // MARK: - Properties
